Move archiver to dedicated namespace.
