+++ UploadDaemon_Test/Archiving/ArchiverTest.cs
@@ -3,55 +3,57 @@ using System.Collections.Generic;
 using System.IO;
 using System.IO.Abstractions;
 using System.IO.Abstractions.TestingHelpers;
-using UploadDaemon;
 
-[TestFixture]
-public class ArchiverTest
+namespace UploadDaemon.Archiving
 {
-    private const string TraceDirectory = @"C:\users\public\traces";
-
-    [Test]
-    public void ShouldMoveFilesToCorrectSubfolders()
+    [TestFixture]
+    public class ArchiverTest
     {
-        IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
-    {
-        { FileInTraceDirectory("coverage_1_1.txt"), @"uploaded" },
-        { FileInTraceDirectory("coverage_1_2.txt"), @"missing version" },
-        { FileInTraceDirectory("coverage_1_3.txt"), @"empty trace" },
-    });
+        private const string TraceDirectory = @"C:\users\public\traces";
 
-        new Archiver(TraceDirectory, fileSystem).ArchiveUploadedFile(FileInTraceDirectory("coverage_1_1.txt"));
-        new Archiver(TraceDirectory, fileSystem).ArchiveFileWithoutVersionAssembly(FileInTraceDirectory("coverage_1_2.txt"));
-        new Archiver(TraceDirectory, fileSystem).ArchiveEmptyFile(FileInTraceDirectory("coverage_1_3.txt"));
-
-        string[] files = fileSystem.Directory.GetFiles(TraceDirectory, "*.txt", SearchOption.AllDirectories);
-
-        Assert.That(files, Is.EquivalentTo(new string[] {
-        FileInTraceDirectory(@"uploaded\coverage_1_1.txt"),
-        FileInTraceDirectory(@"missing-version\coverage_1_2.txt"),
-        FileInTraceDirectory(@"empty-traces\coverage_1_3.txt"),
-    }));
-    }
-
-    [Test]
-    public void ShouldHandleExceptionsGracefully()
-    {
-        IFileSystem fileSystemMock = FileSystemMockingUtils.MockFileSystem(fileMock =>
+        [Test]
+        public void ShouldMoveFilesToCorrectSubfolders()
         {
-            fileMock.Setup(file => file.ReadAllLines(FileInTraceDirectory("coverage_1_1.txt"))).Throws<IOException>();
-        }, directoryMock =>
+            IFileSystem fileSystem = new MockFileSystem(new Dictionary<string, MockFileData>()
         {
-        // not needed
-    }).Object;
+            { FileInTraceDirectory("coverage_1_1.txt"), @"uploaded" },
+            { FileInTraceDirectory("coverage_1_2.txt"), @"missing version" },
+            { FileInTraceDirectory("coverage_1_3.txt"), @"empty trace" },
+        });
 
-        new Archiver(TraceDirectory, fileSystemMock).ArchiveUploadedFile(FileInTraceDirectory("coverage_1_1.txt"));
-    }
+            new Archiver(TraceDirectory, fileSystem).ArchiveUploadedFile(FileInTraceDirectory("coverage_1_1.txt"));
+            new Archiver(TraceDirectory, fileSystem).ArchiveFileWithoutVersionAssembly(FileInTraceDirectory("coverage_1_2.txt"));
+            new Archiver(TraceDirectory, fileSystem).ArchiveEmptyFile(FileInTraceDirectory("coverage_1_3.txt"));
 
-    /// <summary>
-    /// Returns a file with the given name in the trace directory.
-    /// </summary>
-    private string FileInTraceDirectory(string fileName)
-    {
-        return Path.Combine(TraceDirectory, fileName);
+            string[] files = fileSystem.Directory.GetFiles(TraceDirectory, "*.txt", SearchOption.AllDirectories);
+
+            Assert.That(files, Is.EquivalentTo(new string[] {
+            FileInTraceDirectory(@"uploaded\coverage_1_1.txt"),
+            FileInTraceDirectory(@"missing-version\coverage_1_2.txt"),
+            FileInTraceDirectory(@"empty-traces\coverage_1_3.txt"),
+        }));
+        }
+
+        [Test]
+        public void ShouldHandleExceptionsGracefully()
+        {
+            IFileSystem fileSystemMock = FileSystemMockingUtils.MockFileSystem(fileMock =>
+            {
+                fileMock.Setup(file => file.ReadAllLines(FileInTraceDirectory("coverage_1_1.txt"))).Throws<IOException>();
+            }, directoryMock =>
+            {
+            // not needed
+        }).Object;
+
+            new Archiver(TraceDirectory, fileSystemMock).ArchiveUploadedFile(FileInTraceDirectory("coverage_1_1.txt"));
+        }
+
+        /// <summary>
+        /// Returns a file with the given name in the trace directory.
+        /// </summary>
+        private string FileInTraceDirectory(string fileName)
+        {
+            return Path.Combine(TraceDirectory, fileName);
+        }
     }
 }
